Implementaciones

+++ src/main/java/org/prueba/repositorios/IProductoRepository.java
@@ -3,6 +3,6 @@ package org.prueba.repositorios;
 import org.prueba.modelos.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IProductoRepository extends JpaRepository<Producto, Long> {
+public interface IProductoRepository extends JpaRepository<Producto, Integer> {
 
 }
